Hid status bar
The status bar should now be hidden, but I am still working on trying to make the webview element entirely full screen. Such as removing the bottom action buttons on android and adding support for phones with camera cutouts, so far the code is nowhere near, it's just working. RIP