Version 1.0.1 - Preview APK
This is a preview APK in an attempt to fix the download/import support. There are still many things that need to be worked on, as the APK will crash when somebody tries to download something from the Novel AI website, but it can for some reason download PDFs, appears to be a file type handling problem, but I am unsure.

Never the less, this is the current progress of things.

Changelogs

Version 1.0.1

+ Added hardware acceleration on the application level.
+ Added external storage, read and write permissions for the download and export of stories.

- Removed unsued colour values in the file "colors.xml". We are only using the "nai_theme" as the default colour scheme.

Issues

- The program is able to downlaod files such as PDFs via the webview element, but will crash when you click on the downlaod button for the stories inside of the Novel AI website. This may be due to the file type. Tested on the Pixel 3A API 30 x86.

Credit: Sabith Pkc Mnr
(https://www.youtube.com/watch?v=n-iXcR7MHgw)